Move test_st_legacy_dataframe test
